app-backend/pyproject.toml
[tool.poetry]
name = "app-backend"
version = "0.1.0"
description = ""
authors = ["vishnuvyas sethumadhavan <vishnuvyas@gmail.com>"]
readme = "README.md"
packages = [{include = "app_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.2"
pandas = "^2.1.4"
scipy = "^1.11.4"
scikit-learn = "^1.3.2"
matplotlib = "^3.8.2"
statsmodels = "^0.14.0"
cvxopt = "^1.3.2"
pydantic = "^2.5.2"
fastapi = "^0.105.0"
sqlmodel = "^0.0.14"
alembic = "^1.13.0"
pymysql = "^1.1.0"
typer = "^0.9.0"
structlog = "^23.2.0"
uvicorn = "^0.24.0.post1"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
sqlalchemy-utils = "^0.41.1"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-mypy = "^0.10.3"
pytest-cov = "^4.1.0"
typeguard = "^4.1.5"
flake8 = "3.7.9"
pytest-pylint = "^0.21.0"

[tool.pytest.ini_options]
minversion="7.0"
addopts="-ra --mypy --typeguard-packages=app_backend --cov=app_backend --import-mode=importlib"
testpaths=["tests"]

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
